rent_site/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.3
itsdangerous==2.2.0
